Move KilometrikisaSession class to own file
So it does not clutter the index.ts
